requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
Django==1.11.6
django-cors-headers==2.1.0
django-environ==0.4.4
django-grappelli==2.10.1
django-material==1.1.1
django-rest-framework==0.1.0
django-templated-mail==1.0.0
djangorestframework==3.7.1
djangorestframework-jwt==1.11.0
djoser==1.0.1
idna==2.6
kombu==4.1.0
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.4
PyJWT==1.5.3
pytz==2017.2
requests==2.18.4
six==1.11.0
urllib3==1.22
vine==1.1.4
